разделяет сценарий просмотра фотографий
разделяет сценарий просмотра фотографий на два модуля: rendering-photo-fullsize.js (именованный ранее как photo-fullsize-view.js) и photo-modal.js
